prototype for vector issue #2: separate pmwebd & pmcd connection config
This prototype enables a deployment model for vector where a pmwebd
instance serves as a proxy/relay to get to pmcds running on actual
target machines.  This requires the simple host:port/hostname concept
to be split up into three:

a) $rootScope.properties.host     -- vector-to-pmwebd connectivity
b) $rootScope.properties.pmcd     -- pmwebd-to-pmcd connectivity
c) $rootScope.properties.hostname -- the pmcd.hostname from the target

The first two may be specified by a new "accordion" type gui from
angular/bootstrap, and default to the server whence the webapp is
served and local: respectively.  The latter is auto-computed during
context setup, and is the field shown on the normal live dashboard.